+++ src/tools/kdb/cmdline.hpp
@@ -51,6 +51,10 @@ public:
 	bool verbose; /*!< Be more verbose. */
 	bool version; /*!< Return version info instead of the normal action.. */
 	bool withoutElektra;
+	bool null;
+	bool first;
+	bool second;
+	bool third;
 
 	/** The path to the kdb exectuable. */
 	std::string executable;