+++ src/tools/kdb/mountbase.hpp
@@ -42,7 +42,7 @@ class MountBaseCommand : public Command
 {
 
 protected:
-	void readMountConf();
+	void readMountConf(Cmdline const& cl);
 	void fixRootKey(Cmdline const& cl);
 	void getName(Cmdline const& cl);
 	void getMountpoint(Cmdline const& cl);